+++ src/prodWs/Service.ts
@@ -3,6 +3,7 @@ import { WSSecurity } from 'soap';
 import { TYPES } from '../di/types';
 import { DataStore } from '../services/DataStore';
 import { SoapService } from '../services/Soap';
+import { Catalog, parseCatalog } from './catalog';
 import { Client, ClientCredentials } from './Client';
 import { formatDate } from './date';
 import { ClientError } from './Error';
@@ -22,6 +23,8 @@ export interface ProductServiceOptions {
 }
 
 export interface ProdWS {
+  getCatalogList(): Promise<Catalog[]>;
+  getCatalog(name: string): Promise<Catalog | null>;
   getProductList(date?: Date): Promise<Product[]>;
   getProduct(id: number): Promise<Product | null>;
 }
@@ -37,6 +40,7 @@ export class ProductService extends SoapService implements ProdWS {
 
   constructor(
     @inject(TYPES.Client) private client: Client,
+    @inject(TYPES.CatalogStore) private catalogStore: DataStore<Catalog>,
     @inject(TYPES.ProductStore) private productStore: DataStore<Product>,
     @inject(TYPES.LoggerFactory) getLogger: any,
     @inject(TYPES.SoapClientFactory) getSoapClient: any
@@ -57,15 +61,34 @@ export class ProductService extends SoapService implements ProdWS {
       throw new ClientError('Missing client credentials for ProdWS service init.');
     }
 
+    const ttl = undefined !== options.ttl ? options.ttl : 7 * 24 * 3600;
+
     this.client.setCredentials(options.client);
 
     await this.checkSoapClient();
 
-    await this.productStore.init(
-      'product-list.json',
-      this.updateProducts.bind(this),
-      undefined !== options.ttl ? options.ttl : 7 * 24 * 3600
-    );
+    await this.catalogStore.init('catalog-list.json', this.updateCatalogs.bind(this), ttl);
+    await this.productStore.init('product-list.json', this.updateProducts.bind(this), ttl);
+  }
+
+  /**
+   * Retrieves the list of available catalogs from the service.
+   */
+  public async getCatalogList(): Promise<Catalog[]> {
+    await this.checkServiceInit('Cannot get catalog list before initializing product service');
+
+    return this.catalogStore.getList();
+  }
+
+  /**
+   * Retrieves the catalog with the given name if existing.
+   *
+   * @param id The name of the catalog that should be retrieved.
+   */
+  public async getCatalog(name: string): Promise<Catalog | null> {
+    await this.checkServiceInit('Cannot get catalog before initializing product service');
+
+    return this.catalogStore.getItem(name);
   }
 
   /**
@@ -93,6 +116,51 @@ export class ProductService extends SoapService implements ProdWS {
     await this.checkServiceInit('Cannot get product before initializing product service');
 
     return this.productStore.getItem(id);
+  }
+
+  protected initSoapClient(): void {
+    const options = {
+      hasTimeStamp: false,
+      hasTokenCreated: false
+    };
+
+    const security = new WSSecurity(this.client.getUsername(), this.client.getPassword(), options);
+
+    this.soapClient.setSecurity(security);
+  }
+
+  private async updateCatalogs(): Promise<{ [id: number]: any }> {
+    const payload: any = {
+      mandantID: this.client.getId()
+    };
+
+    // pay.load.subMandatId
+
+    payload.catalogProperties = true;
+
+    return this.soapClient
+      .getCatalogListAsync(payload)
+      .then(([response]: any) => {
+        if (!response || 'true' !== response.attributes.success) {
+          return {};
+        }
+
+        const catalogs: { [id: number]: Catalog } = {};
+
+        response.Response.catalogList.catalog.forEach((data: any) => {
+          const catalog = parseCatalog(data);
+
+          if (catalog) {
+            catalogs[catalog.id] = catalog;
+          }
+        });
+
+        return catalogs;
+      })
+      .catch((e: any) => {
+        this.log('getCatalogList', e.root?.Envelope.Body.Fault || e);
+        // throw new SoapError(e.root.Envelope.Body.Fault.faultstring);
+      });
   }
 
   private async updateProducts(date?: Date): Promise<{ [id: number]: Product }> {
@@ -132,16 +200,5 @@ export class ProductService extends SoapService implements ProdWS {
         this.log('getProductList', e.root?.Envelope.Body.Fault || e);
         // throw new SoapError(e.root.Envelope.Body.Fault.faultstring);
       });
-  }
-
-  protected initSoapClient(): void {
-    const options = {
-      hasTimeStamp: false,
-      hasTokenCreated: false
-    };
-
-    const security = new WSSecurity(this.client.getUsername(), this.client.getPassword(), options);
-
-    this.soapClient.setSecurity(security);
   }
 }